Conflicts:
	src/main/java/BrowseData.java
	src/main/java/com/bionic/friendsphotos/dao/PhotoDao.java
	src/main/java/com/bionic/friendsphotos/service/PhotoService.java

+++ src/main/java/com/bionic/friendsphotos/service/DevicesService.java
@@ -5,37 +5,99 @@ import com.bionic.friendsphotos.dao.GroupDao;
 import com.bionic.friendsphotos.entity.Device;
 import com.bionic.friendsphotos.entity.Group;
 import org.apache.commons.lang.StringUtils;
+import org.springframework.transaction.annotation.Transactional;
 
+import javax.inject.Inject;
+import javax.inject.Named;
 import java.math.BigInteger;
 import java.util.*;
 
 /**
  * Created by Bogdan Sliptsov on 17.07.2015.
  */
+
+@Named
+@Transactional
 public class DevicesService {
-    DeviceDao devicesDao;
-    GroupDao groupDao;
+
+    @Inject
+    private DeviceDao devicesDao;
+
+    @Inject
+    private GroupDao groupDao;
 
     public DevicesService() {
-        devicesDao  = new DeviceDao();
-        groupDao = new GroupDao();
     }
 
-    public void addNew(Device device) {
-        devicesDao.create(device);
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to add new device to table 'devices' in DB.
+     *
+     * @param device Device object.
+     * @return 1. Null If device reference equals null.
+     *         2. Created object.
+     */
+    public String addNew(Device device) {
+        if (device == null) return null;
+        return devicesDao.create(device);
     }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to find device by device identifier.
+     *
+     * @param id Unique device identifier.
+     *
+     * @return 1. Null If device doesn't exist.
+     *         2. Sought-for device object.
+     */
     public Device findById(String id) {
+        if (StringUtils.isEmpty(id)) return null;
         return devicesDao.read(id);
     }
 
-    public Device updateDevice(Device obj) { return devicesDao.update(obj);}
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to update device it exists.
+     *
+     * @param device Device object
+     * @return 1. Null If device doesn't exist.
+     *         2. Updated device object.
+     */
+    public Device updateDevice(Device device) {
+        if (device == null) return null;
+        return devicesDao.update(device);
+    }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to delete device from table in DB if it exists.
+     *
+     * @param idDevice Unique device identifier.
+     */
     public void deleteDevice(String idDevice) {
+        if (StringUtils.isEmpty(idDevice)) return;
+        if (devicesDao.read(idDevice) == null) return;
         devicesDao.delete(idDevice);
     }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to return id of all groups in requested device.
+     *
+     * @param devicesId Unique device identifier.
+     *
+     * @return 1. Null Device doesn't exist.
+     *         2. List of id of all groups.
+     */
     public List<Long> getIdOfAllGroupsInCurrentDevice(String devicesId) {
+        if (StringUtils.isEmpty(devicesId)) return null;
+        if (devicesDao.read(devicesId) == null) return null;
         List<Long> list = new ArrayList<>();
         for (Group g: devicesDao.getAllGroupsOfCurrentDevice(devicesId)) {
             list.add(g.getId());
@@ -43,15 +105,44 @@ public class DevicesService {
         return list;
     }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to return all devices in table 'devices' in DB.
+     *
+     * @return List of all devices.
+     */
     public List<Device> getAll() {
         return devicesDao.getAll();
     }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to delete group(if it exists) from requested device(if it exists).
+     *
+     * @param deviceId Unique device identifier.
+     * @param groupId Unique group identifier.
+     */
     public void deleteGroupFromDevice(String deviceId, Long groupId) {
-            devicesDao.deleteGroupFromDevice(deviceId, groupId);
+        if (StringUtils.isEmpty(deviceId) || groupId == null) return;
+        if (devicesDao.read(deviceId) == null) return;
+        if (!devicesDao.read(deviceId).getGroups().contains(groupDao.read(groupId))) return;
+        devicesDao.deleteGroupFromDevice(deviceId, groupId);
     }
 
+    /**
+     * @author Bogdan Sliptsov
+     *
+     * This method is used to add group(if it exists) to requeset device(if it exists).
+     *
+     * @param deviceId Unique device identifier.
+     * @param groupId Unique group identifier
+     */
     public void addGroupToDevice(String deviceId, Long groupId) {
+        if (StringUtils.isEmpty(deviceId) || groupId == null) return;
+        if (devicesDao.read(deviceId) == null) return;
+        if (devicesDao.read(deviceId).getGroups().contains(groupDao.read(groupId))) return;
         devicesDao.addGroupToDevice(deviceId, groupDao.read(groupId));
     }
 
@@ -67,6 +158,7 @@ public class DevicesService {
      */
 
     public Map<String, Object> getCurrentGroupAndUserName(String deviceId) {
+        if (StringUtils.isEmpty(deviceId)) return null;
         Device device = findById(deviceId);
         if (device != null) {
             Map<String, Object> map = new TreeMap<>();
@@ -92,10 +184,7 @@ public class DevicesService {
      */
 
     public boolean loginByFacebook(String deviceId, BigInteger fbProfile, String userName, String description) {
-        if (deviceId == null) return false;
         if (fbProfile == null) return false;
-        if (userName == null) return false;
-        if (description == null) return false;
         if (StringUtils.isEmpty(deviceId)) return false;
         if (StringUtils.isEmpty(userName)) return false;
         if (StringUtils.isEmpty(description)) return false;
@@ -123,14 +212,14 @@ public class DevicesService {
      *
      * @param requestedFbProfileList List of all friends of current user.
      *
-     * @return null 1. If input reference is null;
+     * @return Null 1. If input reference is null;
      *              2. If input list is empty;
      *              3. If nobody in list of friends has FriendPhoto application on his device.
-     * @return Map of all friends that has FriendsPhoto application and also list of their groups.
+     *         Map of all friends that has FriendsPhoto application and also list of their groups.
      */
     public Map<BigInteger, List<Group>> getAllFriendsOfCurrentUser(List<BigInteger> requestedFbProfileList) {
         if (requestedFbProfileList == null) return null;
-        if (requestedFbProfileList.size() == 0) return null;
+        if (requestedFbProfileList.isEmpty()) return null;
         List<Device> deviceList = devicesDao.getAll();
         List<Device> mergedList = new ArrayList<>();
         for (int i = 0; i < deviceList.size(); i++) {
@@ -140,7 +229,7 @@ public class DevicesService {
                 }
             }
         }
-        if (mergedList.size() == 0) return null;
+        if (mergedList.isEmpty()) return null;
         Map<BigInteger, List<Group>> friendsMap = new TreeMap<>();
         for (Device d: mergedList) {
             friendsMap.put(d.getFbProfile(), d.getGroups());
@@ -158,16 +247,15 @@ public class DevicesService {
      * @param deviceId Device unique inedtifier.
      * @param groupId Group unique identifier.
      *
-     * @return false 1. If deviceId reference is null;
+     * @return False 1. If deviceId reference is null;
      *               2. If deficeId is empty;
      *               3. If groupId reference is null;
      *               4. If group doesn't exist;
      *               5. If device doesn't exist;
-     * @return true  1. If group that user want to connect is already current for his device;
+     *         True  1. If group that user want to connect is already current for his device;
      *               2. If current group has been changed.
      */
     public boolean connectUserToOpenGroup(String deviceId, Long groupId) {
-        if (deviceId == null) return false;
         if (StringUtils.isEmpty(deviceId)) return false;
         if (groupId == null) return false;
         Group group = groupDao.read(groupId);
@@ -195,7 +283,7 @@ public class DevicesService {
      * @param groupId Group unique identifier.
      * @param password Password to closed group.
      *
-     * @return false 1. If deviceId reference is null;
+     * @return False 1. If deviceId reference is null;
      *               2. If groupId reference is null;
      *               3. If password reference is null;
      *               4. If deviceId is empty;
@@ -203,13 +291,11 @@ public class DevicesService {
      *               6. If group doesn't exist;
      *               7. If device doesn't exist;
      *               8. If password is incorrect.
-     * @return true  1. If group that user want to connect is already current for his device;
+     *         True  1. If group that user want to connect is already current for his device;
      *               2. If password is correct and current group has been changed.
      */
     public boolean connectUserToClosedGroup(String deviceId, Long groupId, String password) {
-        if (deviceId == null) return false;
         if (groupId == null) return false;
-        if (password == null) return false;
         if (StringUtils.isEmpty(deviceId)) return false;
         if (StringUtils.isEmpty(password)) return false;
         Group group = groupDao.read(groupId);
@@ -233,9 +319,9 @@ public class DevicesService {
      *
      * @param deviceId Device unique inedtifier.
      *
-     * @return null 1. If deviceId reference is null;
+     * @return Null 1. If deviceId reference is null;
      *              2. If device doesn't exist.
-     * @return List of members of users current group.
+     *         List of members of users current group.
      */
     public List<Device> membersOfUsersCurrentGroup(String deviceId) {
         if (deviceId == null) return null;
@@ -253,7 +339,7 @@ public class DevicesService {
      * @param deviceId Device unique inedtifier.
      * @param newGroupName New name for users current group.
      *
-     * @return false 1. If deviceId reference is null;
+     * @return False 1. If deviceId reference is null;
      *               2. If newGroupName reference is null;
      *               3. If deviceId is empty;
      *               4. If newGroupName is empty;
@@ -261,11 +347,9 @@ public class DevicesService {
      *               6. If this device is not a creator of current group;
      *               7. If group doesn't exist;
      *               8. If name wasn't changed.
-     * @return true If name was changed.
+     *         True If name was changed.
      */
     public boolean changeGroupName(String deviceId, String newGroupName) {
-        if (deviceId == null) return false;
-        if (newGroupName == null) return false;
         if (StringUtils.isEmpty(deviceId)) return false;
         if (StringUtils.isEmpty(newGroupName)) return false;
         Device device = findById(deviceId);
@@ -291,7 +375,7 @@ public class DevicesService {
      * @param adminId Device unique identifier which has admin rights.
      * @param idDeviceToRemove Device unique identifier which should be removed.
      *
-     * @return false 1. If adminId reference is null;
+     * @return False 1. If adminId reference is null;
      *               2. If idDeviceToRemove reference is null;
      *               3. If adminId is empty;
      *               4. If idDeviceToRemove is empty;
@@ -300,11 +384,9 @@ public class DevicesService {
      *               7. If group doesn't exist;
      *               8. If adminDevice is not a creator of current group;
      *               9. If device wasn't removed;
-     * @return true If device was removed.
+     *         True If device was removed.
      */
     public boolean removeMembersFromCurrentGroup(String adminId, String idDeviceToRemove) {
-        if (adminId == null) return false;
-        if (idDeviceToRemove == null) return false;
         if (StringUtils.isEmpty(adminId)) return false;
         if (StringUtils.isEmpty(idDeviceToRemove)) return false;
         Device adminDevice = findById(adminId);
